config: Remove config.Loader
After the extraction of the environment variable-related logic
from Loader.Load(), there was no justification to keep the
config.Loader type.
